fix format precio for CarreraTaxi

+++ app/Application/CarreraTaxi/Dto/Command/CreateCarreraTaxiCommand.php
@@ -25,11 +25,11 @@ class CreateCarreraTaxiCommand
     public readonly string $cliente,
     public readonly string $taxi,
     public readonly string $taxista,
-    public readonly float $kilometros,
+    public readonly int $kilometros,
     public readonly string $barrioInicio,
     public readonly string $barrioLlegada,
     public readonly int $cantidadPasajeros,
-    public readonly float $precio,
+    public readonly int $precio,
     public readonly int $duracionMinutos
 ) {
     $this->validate();
@@ -86,16 +86,16 @@ class CreateCarreraTaxiCommand
      */
     public function toArray(): array
     {
-        return [
-          'cliente' => $this->cliente,
-          'taxi' => $this->taxi,
-          'taxista' => $this->taxista,
-          'kilometros' => $this->kilometros,
-          'barrioInicio' => $this->barrioInicio,
-          'barrioLlegada' => $this->barrioLlegada,
-          'cantidadPasajeros' => $this->cantidadPasajeros,
-          'precio' => $this->precio,
-          'duracionMinutos' => $this->duracionMinutos,
-      ];
+      return [
+        'cliente' => $this->cliente,
+        'taxi' => $this->taxi,
+        'taxista' => $this->taxista,
+        'kilometros' => $this->kilometros,
+        'barrioInicio' => $this->barrioInicio,
+        'barrioLlegada' => $this->barrioLlegada,
+        'cantidadPasajeros' => $this->cantidadPasajeros,
+        'precio' => $this->precio,
+        'duracionMinutos' => $this->duracionMinutos,
+    ];
     }
 }
